防止分類信息無法取得緩存的 BUG

+++ discuzx/upload/extensions/hooks/hooks_core.php
@@ -171,6 +171,8 @@ function _eFunc_cachedata_Before_get_syscache($_EVENT, $conf) {
 			if(!isset($_del_cache[$k])
 				// bugfix 修正造成插件語言包無法緩存的問題
 				&& substr($k, 0, 6) != 'plugin'
+				// 防止分類信息無法取得緩存的 BUG
+				&& substr($k, 0, strlen('threadsort')) != 'threadsort'
 			) {
 				DB::query("DELETE FROM ".DB::table('common_syscache')." WHERE cname = '$k' LIMIT 1");
 			}
